Increased mounting point of entities
Implemented  https://github.com/eldoriarpg/PickMeUp/issues/7
However this comes with a drawback. In order to achieve this, I'm adding an invisible baby armor stand as a buffer between the player and the entitiy it is holding, this comes with the drawback of obstructing the player's ability to interact/attack any other block or entity.
The only ones capable of circumbenting this naturally are bedrock player's or modded clients.

To the moment, the only way I've found to fix this is to teleport the armorstand on top of the player every tick instead, however, this makes the entities lag behind when running.